+++ src/main/java/tr/com/milia/resurgence/task/TaskSucceedResult.java
@@ -10,17 +10,17 @@ import java.util.Set;
 
 public class TaskSucceedResult extends TaskResult {
 	private final int experienceGain;
-	private final int moneyGain;
+	private final long moneyGain;
 	private final Set<Skill> skillGain;
 	private final Set<Drop> drop;
 
 	public TaskSucceedResult(Player player,
 							 Task task,
 							 int experienceGain,
-							 int moneyGain,
+							 long moneyGain,
 							 Set<Skill> skillGain,
 							 Set<Drop> drop,
-							 Map<Item, Integer> usedItems) {
+							 Map<Item, Long> usedItems) {
 		super(player, task, usedItems);
 		this.experienceGain = experienceGain;
 		this.moneyGain = moneyGain;
@@ -32,7 +32,7 @@ public class TaskSucceedResult extends TaskResult {
 		return experienceGain;
 	}
 
-	public int getMoneyGain() {
+	public long getMoneyGain() {
 		return moneyGain;
 	}
